Changes made in this release: Sandcastle: - Moved the MemberName and HashedMemberName naming method code into a new AddFilenames.xsl transformation file along with the GUID naming method code. - Deleted AddGuidFilenames.xls and AddFriendlyFilenames.xsl transformations as their functionality is now handled by AddFilenames.xsl. - Breaking Change: The MemberName naming method differs from the old Sandcastle "friendly names" method.  This brings it in line with the SHFB has always worked.  It will only affect people that used to Sandcastle tools directly and only then if they relied on direct links to the previously generated names.  Use the transformation argument /arg:namingMethod=MemberName or /arg:namingMethod=HashedMemberName to use the new naming methods.  The default if not specified is Guid. - Applied a patch from gfraiteur to remove the GUID topic ID requirement in ResolveConceptualLinksComponent.  Bear in mind that GUIDs are still preferred as they are guaranteed to be unique which is important for Help 2 and MS Help Viewer content which are merged into a collection.  Duplicate IDs across multiple sets of content would cause linking issues in the collections. - Added changes from Stazzz to the Version Builder tool to merge information about additional extension methods even when the type and method are defined in different assemblies. - Added support for the .NETFramework 4.5.1, .NETCore 4.5.1, and .NETPortable 4.6 frameworks.
Sandcastle Help File Builder:
- Fixed the determination of the default topic so that it searches for a file in nested topics when the parent topic is an empty container node.
- Fixed documentation source project file handling so that it favors the TargetName property over the AssemblyName property so that it works with C++ projects that have a different assembly name from the project filename.
- Fixed documentation source project file handling to account for the GenerateProjectSpecificOutputFolder property being set to true which modifies how the OutDir property is set for the build.
- Fixed the Assembly Binding Redirection plug-in so that it fully qualifies redirection configuration filenames when written to MRefBuilder.config.
- Fixed the Additional Reference Links and Version Builder plug-ins so that they correctly clone the subprojects to avoid modifying the underlying MSBuild project which may be loaded in Visual Studio.  This prevents the changes they make to the subprojects from propagating to the Visual Studio copy.
- Adjusted how child processes are killed when a build is aborted to try and prevent killing unrelated processes from a separate, concurrent build.
- Enabled window selection commands for use during a build in the standalone GUI.  Also enabled switching between project property categories in the Properties window during a build to allow seeing what values are in effect.
- Removed the ModifyHelpTopicFilenames build step as naming is now handled entirely by the new AddFilenames.xsl file in Sandcastle.

+++ SHFB/Source/SandcastleBuilderUtils/BuildEngine/BuildProcess.cs
@@ -2,7 +2,7 @@
 // System  : Sandcastle Help File Builder Utilities
 // File    : BuildProcess.cs
 // Author  : Eric Woodruff  (Eric@EWoodruff.us)
-// Updated : 07/17/2013
+// Updated : 10/12/2013
 // Note    : Copyright 2006-2013, Eric Woodruff, All rights reserved
 // Compiler: Microsoft Visual C#
 //
@@ -43,15 +43,16 @@
 //                           honor solution-level per-project configuration and platform settings.  Added
 //                           support for multi-format build output. Moved GenerateIntermediateTableOfContents
 //                           so that it occurs right after MergeTablesOfContents.
-#endregion
 // 1.9.1.0  07/09/2010  EFW  Updated for use with .NET 4.0 and MSBuild 4.0.
 // 1.9.2.0  01/16/2011  EFW  Updated to support selection of Silverlight Framework versions
 // 1.9.3.2  08/20/2011  EFW  Updated to support selection of .NET Portable Framework versions
+#endregion
 // 1.9.4.0  03/25/2012  EFW  Merged changes for VS2010 style from Don Fehr
 // 1.9.5.0  09/10/2012  EFW  Updated to use the new framework definition file for the .NET Framework versions
 // 1.9.6.0  10/25/2012  EFW  Updated to use the new presentation style definition files
 // 1.9.7.0  01/02/2013  EFW  Added method to get referenced namespaces
-// 1.9.8.0  06/21/2013  EFW  Added support for format-specific help content files
+// 1.9.8.0  06/21/2013  EFW  Added support for format-specific help content files.  Removed the
+//                           ModifyHelpTopicFilenames build step.
 //===============================================================================================================
 
 using System;
@@ -443,8 +444,8 @@ namespace SandcastleBuilder.Utils.BuildEngine
         }
 
         /// <summary>
-        /// This read-only property returns a hash set used to contain a list of namespaces referenced by
-        /// the reflection data files.
+        /// This read-only property returns a hash set used to contain a list of namespaces referenced by the
+        /// project reflection data files, project XML comments files, and base framework XML comments files.
         /// </summary>
         /// <value>These namespaces are used to limit what the Resolve Reference Links component has to index</value>
         public HashSet<string> ReferencedNamespaces
@@ -1019,13 +1020,6 @@ namespace SandcastleBuilder.Utils.BuildEngine
                 reflectionInfo.Load(reflectionFile);
                 apisNode = reflectionInfo.SelectSingleNode("reflection/apis");
 
-                // Alter the help topic filenames if necessary
-                this.ModifyHelpTopicFilenames();
-
-                // Backup the original reflection file for reference and save the changed file
-                File.Copy(reflectionFile, Path.ChangeExtension(reflectionFile, ".bak"), true);
-
-                reflectionInfo.Save(reflectionFile);
                 commentsFiles.Save();
 
                 this.EnsureOutputFoldersExist("html");
@@ -1515,17 +1509,40 @@ AllDone:
             }
             catch(ThreadAbortException )
             {
-                // Kill off the process, known child processes and the STDOUT
-                // and STDERR threads too if necessary.
+                // Kill off the process, known child processes, and the STDOUT and STDERR threads too if
+                // necessary.
                 if(currentProcess != null && !currentProcess.HasExited)
                 {
-                    currentProcess.Kill();
+                    DateTime procStart;
+
+                    // Only kill potential matches started after the current process's start time.  It's not
+                    // perfect if you've got two or more SHFB builds running concurrently but it's the best
+                    // we can do without getting really complicated which I'm not prepared to do since this is
+                    // an extremely low occurrence issue.
+                    try
+                    {
+                        procStart = currentProcess.StartTime;
+                        currentProcess.Kill();
+                    }
+                    catch
+                    {
+                        // If we can't get the start time, assume the build start time
+                        procStart = buildStart;
+                    }
 
                     foreach(Process p in Process.GetProcesses())
-                        if(reKillProcess.IsMatch(p.ProcessName) && !p.HasExited)
+                        try
                         {
-                            System.Diagnostics.Debug.WriteLine("Killing " + p.ProcessName);
-                            p.Kill();
+                            if(reKillProcess.IsMatch(p.ProcessName) && !p.HasExited && p.StartTime > procStart)
+                            {
+                                System.Diagnostics.Debug.WriteLine("Killing " + p.ProcessName);
+                            
+                                p.Kill();
+                            }
+                        }
+                        catch
+                        {
+                            // Ignore exceptions, the process had probably already exited
                         }
                 }
 
@@ -1567,8 +1584,7 @@ AllDone:
 
                 } while(ex != null);
 
-                // NOTE: Message may contain format markers so pass it as a
-                // format argument.
+                // NOTE: Message may contain format markers so pass it as a format argument
                 if(bex != null)
                     this.ReportError(BuildStep.Failed, bex.ErrorCode, "{0}", message);
                 else
@@ -2000,7 +2016,7 @@ AllDone:
             Version fileVersion = new Version(fvi.FileMajorPart, fvi.FileMinorPart, fvi.FileBuildPart,
                 fvi.FilePrivatePart);
 
-            Version expectedVersion = new Version("2.7.3.0");
+            Version expectedVersion = new Version("2.7.4.0");
 
             if(fileVersion < expectedVersion)
                 throw new BuilderException("BE0036", String.Format(CultureInfo.InvariantCulture,
